Users url is ready
